osan 4 loput tehtävät

+++ osa_4/blogilista/controllers/blogs.js
@@ -1,8 +1,13 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
+const User = require("../models/user")
+const jwt = require("jsonwebtoken")
+
 
 blogsRouter.get("/", async (request, response) => {
-	const blogs = await Blog.find({})
+	const blogs = await Blog.find({}).populate(
+		"user", {username: 1, name: 1}
+	)
 	response.json(blogs.map(blog => blog.toJSON()))
 })
 
@@ -22,10 +27,25 @@ blogsRouter.get("/:id", async (request, response, next) => {
 })
 
 blogsRouter.post("/", async (request, response, next) => {
-	const blog = new Blog(request.body)
-
+	const body = request.body
+	console.log(body)
 	try {
+		const decodedToken = jwt.verify(request.token, process.env.SECRET)
+		if (!request.token || !decodedToken.id) {
+			return response.status(401).json({error: "Token missing or invalid"})
+		}
+		const user = await User.findById(decodedToken.id)
+
+		const blog = new Blog({
+			title: body.title,
+			author: body.author,
+			url: body.url,
+			likes: body.likes,
+			user: user._id
+		})
 		const savedBlog = await blog.save()
+		user.blogs = user.blogs.concat(savedBlog._id)
+		await user.save()
 		response.json(savedBlog.toJSON())
 	}
 	catch (error) {
@@ -35,8 +55,21 @@ blogsRouter.post("/", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
 	try {
-		await Blog.findByIdAndRemove(request.params.id)
-		response.status(204).end()
+		const decodedToken = jwt.verify(request.token, process.env.SECRET)
+		if (!request.token || !decodedToken.id) {
+			return response.status(401).json({error: "Token missing or invalid"})
+		}
+		const user = await User.findById(decodedToken.id)
+		const blog = await Blog.findById(request.params.id)
+		if (blog.user.toString() !== user.id.toString()) {
+			return response.status(401).json({
+				error: "User is not authorized to remove this blog"
+			})
+		}
+		else {
+			await Blog.findByIdAndRemove(request.params.id)
+			response.status(204).end()
+		}
 	}
 	catch (error) {
 		next(error)
